add check url params with sort filter
