+++ tests/models.ts
@@ -5,6 +5,14 @@ export type DateMetadataLabel = {
 
 export type LayoutViewMode = 'tile' | 'list' | 'compact';
 
+export enum SearchOption  {
+  METADATA = `Search metadata`,
+  TEXT = `Search text contents`,
+  TV = `Search TV news captions`,
+  RADIO = `Search radio transcripts`,
+  WEB = `Search archived web sites`,
+};
+
 export type SortFilter = 
   | 'Weekly views' 
   | 'All-time views' 
@@ -16,3 +24,14 @@ export type SortFilter =
   | 'Creator';
 
 export type SortOrder = 'ascending' | 'descending';
+
+export const SortFilterURL = {
+  'Weekly views': 'week',
+  'All-time views': 'downloads',
+  'Title': 'title',
+  'Date published': 'date',
+  'Date archived': 'publicdate',
+  'Date reviewed': 'reviewdate',
+  'Date added': 'addeddate',
+  'Creator': 'creator',
+}
